Fixed convoy route computation through ports
Convoys through ports currently only work if the port is the only step of the route, going start-port-destination. Ports cannot be part of longer convoy chains.

The cause if the `elif` line 113, since ports (being valid destinations of convoys) will match the condition for the `if` line 101. However, ports are both possible destinations and possible next water-like regions, so the two test should be independent `if`s.